@@ -1,4 +1,4 @@
-const esbuild = require('esbuild');
+import { context } from 'esbuild';
 
 const production = process.argv.includes('--production');
 const watch = process.argv.includes('--watch');
@@ -26,17 +26,18 @@ const esbuildProblemMatcherPlugin = {
 };
 
 async function main() {
-  const ctx = await esbuild.context({
+  const ctx = await context({
     entryPoints: ['src/extension.ts'],
     bundle: true,
-    format: 'cjs',
+    format: 'esm',
     minify: production,
     sourcemap: !production,
     sourcesContent: false,
+
     platform: 'node',
     outfile: 'dist/extension.js',
     external: ['vscode'],
-    logLevel: 'silent',
+    logLevel: 'debug',
     plugins: [
       /* add to the end of plugins array */
       esbuildProblemMatcherPlugin,
