@@ -1,8 +1,8 @@
-export * from './extension';
-export * from './constants';
-export * from './process';
-export * from './workspace-folder';
-export * from './workspace';
-export * from './mutation-server';
-export * from './errors';
-export * from './test-explorer';
+export * from './extension.ts';
+export * from './constants.ts';
+export * from './process.ts';
+export * from './workspace-folder.ts';
+export * from './workspace.ts';
+export * from './mutation-server.ts';
+export * from './errors.ts';
+export * from './test-explorer.ts';
