@@ -1,13 +1,13 @@
-import { commonTokens, tokens } from './di';
-import { ServerLocation } from './domain';
+import { commonTokens } from './di/index.ts';
+import type { ServerLocation } from './domain/index.ts';
 import * as net from 'net';
-import { ContextualLogger } from './logging';
+import { ContextualLogger } from './logging/index.ts';
 import {
   JSONRPCClient,
   JSONRPCErrorException,
-  JSONRPCRequest,
+  type JSONRPCRequest,
 } from 'json-rpc-2.0';
-import { JsonRpcEventDeserializer } from './utils';
+import { JsonRpcEventDeserializer } from './utils/index.ts';
 import { promisify } from 'util';
 import {
   ConfigureParams,
@@ -18,37 +18,31 @@ import {
   MutationTestResult,
 } from 'mutation-server-protocol';
 import { filter, map, Subject } from 'rxjs';
-
 const rpcMethods = Object.freeze({
   configure: 'configure',
   discover: 'discover',
   mutationTest: 'mutationTest',
   reportMutationTestProgressNotification: 'reportMutationTestProgress',
 });
-
 export class MutationServer {
+  private readonly logger;
   #socket: net.Socket;
   #jsonRPCClient: JSONRPCClient;
   #notifications = new Subject<JSONRPCRequest>();
-
-  public static readonly inject = tokens(
+  public static readonly inject = [
     commonTokens.contextualLogger,
     commonTokens.serverLocation,
-  );
-  constructor(
-    private readonly logger: ContextualLogger,
-    serverLocation: ServerLocation,
-  ) {
+  ] as const;
+  constructor(logger: ContextualLogger, serverLocation: ServerLocation) {
+    this.logger = logger;
     this.#socket = net.connect(serverLocation.port, serverLocation.host, () => {
       this.logger.info('Connected to server');
     });
-
     this.#jsonRPCClient = new JSONRPCClient((jsonRPCRequest) => {
       const content = Buffer.from(JSON.stringify(jsonRPCRequest));
       this.#socket.write(`Content-Length: ${content.byteLength}\r\n\r\n`);
       this.#socket.write(content);
     });
-
     const deserializer = new JsonRpcEventDeserializer();
     this.#socket.on('data', (data) => {
       const events = deserializer.deserialize(data);
@@ -61,7 +55,6 @@ export class MutationServer {
       }
     });
   }
-
   public async configure(
     configureParams: ConfigureParams,
   ): Promise<ConfigureResult> {
@@ -70,7 +63,6 @@ export class MutationServer {
       configureParams,
     );
   }
-
   public async discover(
     discoverParams: DiscoverParams,
   ): Promise<DiscoverResult> {
@@ -79,7 +71,6 @@ export class MutationServer {
       discoverParams,
     );
   }
-
   public async mutationTest(
     mutationTestParams: MutationTestParams,
     onPartialResult: (partialResult: MutationTestResult) => void,
@@ -94,7 +85,6 @@ export class MutationServer {
         map((notification) => notification.params),
       )
       .subscribe(onPartialResult);
-
     try {
       const result = await this.#jsonRPCClient.request(
         rpcMethods.mutationTest,
@@ -107,7 +97,6 @@ export class MutationServer {
       subscription.unsubscribe();
     }
   }
-
   public async dispose() {
     await promisify(this.#socket.end.bind(this.#socket))();
   }
