@@ -1,4 +1,4 @@
-export * from './json-rpc-event-deserializer';
-export * from './location-utils';
-export * from './test-controller-utils';
-export * from './test-item-utils';
+export * from './json-rpc-event-deserializer.ts';
+export * from './location-utils.ts';
+export * from './test-controller-utils.ts';
+export * from './test-item-utils.ts';
